Simplify implementation of the server
