+++ packages/airnode-node/docker/entrypoint.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-node /app/src/cli/validate-config.js && node /app/src/workers/gateways-api.js & crond -f
+node /app/src/cli/validate-config.js && node /app/src/workers/local-gateways/run-server.js & crond -f
