+++ packages/airnode-node/package.json
@@ -15,6 +15,7 @@
     "dev:invoke": "ts-node src/cli/validate-config.ts && ts-node src/cli/invoke.ts",
     "dev:http-data-invoke": "ts-node src/cli/http-data-invoke.ts",
     "dev:http-signed-data-invoke": "ts-node src/cli/http-signed-data-invoke.ts",
+    "dev:run-gateways-server": "ts-node -T src/workers/local-gateways/run-server.ts",
     "pack": "yarn pack",
     "test:e2e": "SILENCE_LOGGER=true jest --selectProjects e2e",
     "test:e2e:debug": "jest \"http.feature\" --selectProjects e2e",
@@ -31,6 +32,7 @@
     "@api3/ois": "^1.1.1",
     "@api3/promise-utils": "^0.3.0",
     "aws-sdk": "^2.1188.0",
+    "body-parser": "^1.20.0",
     "dotenv": "^16.0.1",
     "ethers": "^5.6.9",
     "express": "^4.18.1",
